requirements.txt
torch==2.3.1
torchvision==0.18.1
timm==1.0.3
ultralytics>=8.0.0               

mlflow==2.13.0

opencv-python-headless==4.9.0.80
Pillow==10.3.0
scikit-image==0.23.2
imageio==2.34.1

numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
scikit-learn==1.5.0

PyYAML==6.0.1
python-dotenv==1.0.1
omegaconf==2.3.0

matplotlib==3.9.0
seaborn==0.13.2

tqdm==4.66.4
rich==13.7.1
loguru==0.7.2

jupyterlab==4.2.1
ipywidgets==8.1.3
